refactor: flat structure
