--- lerna.json
@@ -1,18 +1,0 @@
-{
-  "npmClient": "yarn",
-  "useWorkspaces": true,
-  "conventionalCommits": true,
-  "exact": true,
-  "packages": [
-    "packages/*"
-  ],
-  "command": {
-    "publish": {
-      "npmClient": "npm"
-    },
-    "init": {
-      "exact": true
-    }
-  },
-  "version": "independent"
-}
