+++ package.json
@@ -1,31 +1,48 @@
 {
-  "private": true,
-  "workspaces": [
-    "packages/*"
+  "name": "ipx",
+  "version": "0.3.2",
+  "repository": "nuxt-contrib/ipx",
+  "license": "MIT",
+  "main": "dist/ipx.cjs.js",
+  "bin": "bin/ipx.js",
+  "files": [
+    "dist",
+    "bin",
+    "types",
+    "client"
   ],
   "scripts": {
-    "start": "cross-env DEBUG=1 IPX_DIST=../src node packages/ipx/bin/ipx.js",
     "dev": "nodemon",
-    "lint": "eslint --ext .ts .",
-    "build": "lerna run build",
-    "prepublish": "yarn build",
-    "release": "lerna publish --conventional-commits"
+    "start": "node bin/ipx.js"
+  },
+  "dependencies": {
+    "connect": "^3.7.0",
+    "consola": "^2.15.0",
+    "cron": "^1.8.2",
+    "defu": "^3.1.0",
+    "etag": "^1.8.1",
+    "fs-extra": "^9.0.1",
+    "is-valid-path": "^0.1.1",
+    "node-fetch": "^2.6.1",
+    "recursive-readdir": "^2.2.2",
+    "sharp": "^0.26.0",
+    "siroc": "^0.4.0"
   },
   "devDependencies": {
-    "@types/etag": "^1.8.0",
+    "@nuxtjs/eslint-config-typescript": "^3.0.0",
     "@types/connect": "^3.4.33",
     "@types/cron": "^1.7.2",
+    "@types/etag": "^1.8.0",
     "@types/fs-extra": "^9.0.1",
     "@types/is-valid-path": "^0.1.0",
+    "@types/node-fetch": "^2.5.7",
     "@types/recursive-readdir": "^2.2.0",
-    "@types/sharp": "^0.25.1",
-    "@nuxtjs/eslint-config-typescript": "^3.0.0",
-    "bili": "^3.4.2",
-    "cross-env": "^5.2.0",
-    "eslint": "^5.12.1",
-    "lerna": "3.10.7",
-    "nodemon": "^1.18.9",
-    "rollup-plugin-typescript2": "^0.27.2",
-    "typescript": "^4.0.2"
+    "@types/sharp": "^0.26.0",
+    "cross-env": "^7.0.2",
+    "eslint": "^7.10.0",
+    "nodemon": "^2.0.4",
+    "rollup-plugin-typescript2": "^0.27.3",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3"
   }
 }
