--- packages/ipx-client/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "ipx-client",
-  "version": "0.5.2",
-  "repository": "nuxt-contrib/ipx",
-  "license": "MIT",
-  "main": "dist/ipx-client.cjs.js",
-  "module": "dist/ipx-client.es.js",
-  "browser": "dist/ipx-client.js",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "bili --format umd,cjs,es src/index.js"
-  }
-}
