--- packages/ipx/CHANGELOG.md
@@ -1,29 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.3.2](https://github.com/nuxt-contrib/ipx/compare/ipx@0.3.1...ipx@0.3.2) (2020-09-23)
-
-**Note:** Version bump only for package ipx
-
-
-
-
-
-## [0.3.1](https://github.com/jsless/ipx/compare/ipx@0.3.0...ipx@0.3.1) (2019-01-27)
-
-**Note:** Version bump only for package ipx
-
-
-
-
-
-# 0.3.0 (2018-12-17)
-
-
-### Features
-
-* allow port customization ([1bdb84e](https://github.com/jsless/ipx/commit/1bdb84e))
-* refactors and improvements ([f378d6e](https://github.com/jsless/ipx/commit/f378d6e))
-* rewrite server into a pure connect middleware ([cc5b911](https://github.com/jsless/ipx/commit/cc5b911))
