--- packages/ipx/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "ipx",
-  "version": "0.3.2",
-  "repository": "nuxt-contrib/ipx",
-  "license": "MIT",
-  "main": "dist/ipx.cjs.js",
-  "bin": "bin/ipx.js",
-  "files": [
-    "dist",
-    "bin"
-  ],
-  "scripts": {
-    "build": "bili --format cjs src/index.ts"
-  },
-  "dependencies": {
-    "connect": "^3.7.0",
-    "consola": "^2.15.0",
-    "cron": "^1.8.2",
-    "defu": "^3.1.0",
-    "etag": "^1.8.1",
-    "fs-extra": "^9.0.1",
-    "is-valid-path": "^0.1.1",
-    "node-fetch": "^2.6.1",
-    "recursive-readdir": "^2.2.2",
-    "sharp": "^0.26.0"
-  },
-  "devDependencies": {
-    "@types/node-fetch": "^2.5.7"
-  }
-}
